refactor: updated the folder structures of the repo, and deleted innecessary 'main' folders in scripts and utils, also the tasks that used the files in hardhat.config.ts
