+++ hardhat.config.ts
@@ -1,14 +1,11 @@
 import * as dotenv from "dotenv";
 
-import { HardhatUserConfig, task, subtask } from "hardhat/config";
-import { NetworkUserConfig } from "hardhat/types/config";
+import { HardhatUserConfig, task } from "hardhat/config";
 import "@nomiclabs/hardhat-etherscan";
 import "@nomiclabs/hardhat-waffle";
 import "@typechain/hardhat";
 import "hardhat-gas-reporter";
 import "solidity-coverage";
-
-import * as main from "./src/utils/main";
 
 import pino from "pino";
 const logger = pino();
@@ -33,116 +30,6 @@ task("accounts", "Prints the list of accounts", async (taskArgs, hre) => {
     console.log(account.address);
   }
 });
-
-task("dfp-main", "Run main contract tasks")
-  .addPositionalParam("action", "The contract action")
-  .addOptionalParam("mainAddr", "optional: address of the main contract")
-  .addOptionalParam("strategyName", "The name of the strategy")
-  .addOptionalParam(
-    "strategyAddr",
-    "optional: address of the strategy contract"
-  )
-  .setAction(async (taskArgs, { run }) => {
-    switch (taskArgs.action) {
-      case "get-strategies": {
-        await run("dfp-main-get-strategies", { ...taskArgs });
-        break;
-      }
-
-      case "add-strategy": {
-        await run("dfp-main-add-strategy", { ...taskArgs });
-        break;
-      }
-
-      case "delete-strategy": {
-        await run("dfp-main-delete-strategy", { ...taskArgs });
-        break;
-      }
-
-      default:
-        logger.error("invalid main contract action");
-    }
-  });
-
-subtask("dfp-main-get-strategies", "Get strategies of main contract")
-  .addOptionalParam("mainAddr", "The address of main contract")
-  .setAction(async ({ mainAddr }, { ethers }) => {
-    // define main contract
-    const contract = await main.getOrDeployContract(ethers, mainAddr);
-
-    // get strategies
-    const [names, addreses] = await contract.getStrategies();
-    const strategies = names.reduce(
-      (
-        arr: { name: string; address: string }[],
-        name: string,
-        index: number
-      ) => {
-        const address = addreses[index];
-        return [...arr, { name, address }];
-      },
-      []
-    );
-
-    console.log("Strategies: ", strategies);
-  });
-
-subtask("dfp-main-add-strategy", "Add strategy to main contract")
-  .addOptionalParam("mainAddr", "The address of main contract")
-  .addParam("strategyName", "The name of the strategy")
-  .addParam("strategyAddr", "The address of the strategy contract")
-  .setAction(async ({ mainAddr, strategyName, strategyAddr }, { ethers }) => {
-    // define main contract
-    const contract = await main.getOrDeployContract(ethers, mainAddr);
-
-    // create strategy
-    await main.addStrategy(
-      ethers,
-      contract.address,
-      strategyAddr,
-      strategyName
-    );
-    console.log(
-      `Created strategy name=${strategyName} address=${strategyAddr}`
-    );
-  });
-
-subtask("dfp-main-delete-strategy", "Delete strategy to main contract")
-  .addOptionalParam("mainAddr", "The address of main contract")
-  .addParam("strategyAddr", "The address of the strategy contract")
-  .setAction(async ({ mainAddr, strategyAddr }, { ethers }) => {
-    // define main contract
-    const contract = await main.getOrDeployContract(ethers, mainAddr);
-
-    // delete strategy
-    await main.deleteStrategy(ethers, contract.address, strategyAddr);
-    console.log(`Deleted strategy address=${strategyAddr}`);
-  });
-
-task("dfp-strategy", "Run strategy contract tasks")
-  .addPositionalParam("strategyName", "The name of the strategy")
-  .addPositionalParam("action", "The contract action")
-  .addOptionalParam("mainAddr", "optional: address of the main contract")
-  .addOptionalParam(
-    "strategyAddr",
-    "optional: address of the strategy contract"
-  )
-  .setAction(async (taskArgs, { run }) => {
-    switch (taskArgs.action) {
-      case "deposit": {
-        await run("dfp-strategy-deposit", { ...taskArgs });
-        break;
-      }
-
-      case "withdraw": {
-        await run("dfp-strategy-withdraw", { ...taskArgs });
-        break;
-      }
-
-      default:
-        logger.error("invalid strategy contract action");
-    }
-  });
 
 // You need to export an object to set up your config
 // Go to https://hardhat.org/config/ to learn more
