+++ src/contracts/interfaces/IStrategy.sol
