+++ src/scripts/aave/get-account-info.ts
