+++ src/scripts/strat-recursive-farming/deposit.ts
