+++ src/scripts/strat-recursive-farming/index.ts
