+++ src/scripts/strat-supply-aave/deploy-mock.ts
