+++ src/test/strat-recursive-farming/contracts/keepers-internal.test.ts
@@ -5,7 +5,7 @@ import {
   DFP,
   MockPoolDFP,
   IRewardsController__factory,
-} from "../typechain";
+} from "../../../typechain";
 import { expect, use } from "chai";
 import "@nomiclabs/hardhat-ethers";
 import { network, ethers } from "hardhat";
