--- src/utils/example.test.ts
@@ -1,7 +1,0 @@
-import { expect } from "chai";
-
-describe("Example", () => {
-  it("example", async () => {
-    expect(1).to.equal(1);
-  });
-});
