--- src/utils/main/index.ts
@@ -1,5 +1,0 @@
-export * from "./deploy-contract";
-export * from "./get-contract";
-export * from "./get-or-deploy-contract";
-export * from "./add-strategy";
-export * from "./delete-strategy";
