--- src/utils/strategy/index.ts
@@ -1,4 +1,0 @@
-import * as recursiveFarming from "./recursive-farming";
-export * from "./deploy-contract";
-
-export { recursiveFarming };
